+++ Heartland-iOS-SDK-SampleApp/Heartland-iOS-SDK-SampleApp/ConnectC2XViewController.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class ConnectC2XViewController: UIViewController {
     var device: HpsC2xDevice?
+    var paxDevice: HpsPaxDevice?
     private let notificationCenter: NotificationCenter = .default
 
     @IBOutlet var connectionLabel: UILabel!
@@ -24,7 +25,7 @@ class ConnectC2XViewController: UIViewController {
         let config = HpsConnectionConfig()
         config.username = ""
         config.password = ""
-        config.siteID = "";
+        config.siteID = ""
         config.deviceID = ""
         config.licenseID = ""
         config.developerID = ""
@@ -35,11 +36,93 @@ class ConnectC2XViewController: UIViewController {
         device?.deviceDelegate = self
         device?.scan()
         activityIndicator.isHidden = false
-        
-        /// Please, uncommente these lines to test Pax Integration for Card Brand Transaction id + Token.
-    
-//        HpsPaxDeviceTransaction.testPaxDeviceManualWithSecondTransactionUsingToken()
-//        HpsPaxDeviceTransaction.testPaxDeviceManualWithSecondTransactionUsingTokenAuth()
+    }
+
+    func testPaxDeviceManual() {
+        let timeout = 120
+
+        let config = HpsConnectionConfig()
+        config.ipAddress = ""
+        config.port = ""
+        config.username = ""
+        config.password = ""
+        config.siteID = ""
+        config.deviceID = ""
+        config.licenseID = ""
+        config.developerID = ""
+        config.versionNumber = ""
+        config.connectionMode = 1
+        config.timeout = timeout
+
+        paxDevice = HpsPaxDevice(config: config)
+
+        let builder = HpsPaxCreditSaleBuilder(device: paxDevice)
+        builder?.amount = 11.0
+        builder?.referenceNumber = 10
+        builder?.allowDuplicates = false
+
+        builder?.execute { response, error in
+
+            if let error = error {
+                print("Error: \(error)")
+                return
+            }
+            if let response = response {
+                print("Response: \(response)")
+                let responseReturn = response.parseResponse()
+                print("Response Parse: \(responseReturn.debugDescription)")
+            }
+        }
+    }
+
+    func testPaxDeviceAuth() {
+        let timeout = 120
+
+        let config = HpsConnectionConfig()
+        config.ipAddress = ""
+        config.port = ""
+        config.username = ""
+        config.password = ""
+        config.siteID = "";
+        config.deviceID = ""
+        config.licenseID = ""
+        config.developerID = ""
+        config.versionNumber = ""
+        config.connectionMode = 1
+        config.timeout = timeout
+
+        paxDevice = HpsPaxDevice(config: config)
+
+        let card = HpsCreditCard()
+        card.cardNumber = ""
+        card.expMonth = 1
+        card.expYear = 2
+        card.cvv = ""
+
+        let address = HpsAddress()
+        address.address = ""
+        address.zip = ""
+
+        let builder = HpsPaxCreditAuthBuilder(device: paxDevice)
+        builder?.amount = 11.0
+        builder?.referenceNumber = 1
+        builder?.allowDuplicates = true
+        builder?.requestMultiUseToken = true
+        builder?.creditCard = card
+        builder?.address = address
+
+        builder?.execute { response, error in
+
+            if let error = error {
+                print("Error: \(error)")
+                return
+            }
+            if let response = response {
+                print("Response: \(response)")
+                let responseReturn = response.parseResponse()
+                print("Response Parse: \(responseReturn.debugDescription)")
+            }
+        }
     }
 }
 
@@ -96,5 +179,58 @@ private extension ConnectC2XViewController {
     enum UIUserInterfaceIdiom: Int {
         case phone // iPhone and iPod touch style UI
         case pad // iPad style UI (also includes macOS Catalyst)
+    }
+}
+
+
+extension ConnectC2XViewController {
+    public func setupUpaDevice() -> HpsUpaDevice? {
+            let config = HpsConnectionConfig()
+            config.ipAddress = "192.168.31.118" //"192.168.31.107" // 199
+            config.port = "8081"
+            config.connectionMode = HpsConnectionModes.TCP_IP.rawValue
+            config.timeout = 1000
+            return HpsUpaDevice(config: config)
+     }
+    
+    public func testingUPAUsaSignatureData() {
+        let device = setupUpaDevice()
+        if let builder = HpsUpaSaleBuilder(device: device) {
+            builder.ecrId = "13"
+            builder.clerkId = "1234"
+            builder.referenceNumber = 1234
+            builder.amount = 30.0;
+            builder.details = HpsTransactionDetails()
+            builder.details.invoiceNumber = generateInvoiceNumber()
+
+            builder.execute(forUPAUSA: { response, error in
+                if let error = error {
+                    print(" Error: \(error)")
+                    return
+                }
+
+                if let response = response {
+                    print(" Response return ")
+                    print(response.approvalCode)
+                    print(" Device signature data")
+                    device?.getSignatureData(builder.ecrId, andRequestId: "1234", response: { signatureDataResponse, error in
+                        if let error = error {
+                            print(" Error: \(error)")
+                            return
+                        }
+                        
+                        if let signatureDataResponse = signatureDataResponse {
+                            print(" Signature Response")
+                            print(signatureDataResponse.data.data.signatureData)
+                        }
+                    })
+                }
+            })
+        }
+    }
+    
+    func generateInvoiceNumber() -> String {
+        let number = Int.random(in: 0001..<5000)
+        return "\(number)"
     }
 }