+++ Heartland-iOS-SDK/Classes/Terminals/HpsConnectionConfig.m
@@ -3,13 +3,6 @@
 
 @implementation HpsConnectionConfig
 
--(id)init {
-     if (self = [super init])  {
-         self.timeout = 500; //Seconds
-     }
-     return self;
-}
-
 - (void)setLogger:(id<HpsInterfaceLogging>)logger {
     _logger = logger;
     [_logger willLogHPSInterfaceWithConfig:self];