+++ Heartland-iOS-SDK/Classes/Terminals/PAX/Models/HpsPaxCreditResponse.h
@@ -4,7 +4,6 @@
 @interface HpsPaxCreditResponse : HpsPaxDeviceResponse
 
 @property (nonatomic,strong) NSString *authorizationCode;
-@property (nonatomic,strong) NSString *cardBrandTransactionId;
 
 - (id) initWithBuffer:(NSData*)buffer;
 - (HpsBinaryDataScanner*) parseResponse;