+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Models/HpsUpaDeviceSignatureResponse.m
@@ -1,0 +1,18 @@
+//
+//  HpsUpaDeviceSignatureResponse.m
+//  Heartland-iOS-SDK
+//
+
+#import "HpsUpaDeviceSignatureResponse.h"
+
+@implementation HpsUpaDeviceSignatureResponse
+@end
+
+@implementation HpsUpaDeviceSignatureResponseData
+@end
+
+@implementation HpsUpaCmdResult
+@end
+
+@implementation HpsUpaSignatureData
+@end