+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 All notable changes to this project will be documented in this file.
  
 ## 2.0.8 2023-05-23
- 
 ### Added
 - Added Timeout functionality to PayApp TCP connection. (US344949)
 - Added token and card brand transaction id to the response for Pax. (US345636)
+
+## 2.0.9 2023-06-13
+### Added
+- Added UPA Signature Data. (US349346)
+- Added UPA Pay App Canada (US346223)
+- iOS SDK: C2X Refund Reversal (US359118)
+- Bug fixes and improvements.
  
 ### Changed